fix: rename queuemanager to queueclient
